requirements.txt
absl-py==1.1.0
addict==2.4.0
aiofiles==22.1.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.18.0
albumentations==0.4.3
altair==4.2.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
#asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
#backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.zoneinfo==0.2.1
basicsr==1.3.5
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
blobfile==1.3.1
boto3==1.24.11
botocore==1.27.11
braceexpand==0.1.7
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@b46f5ac7587d2e1862f8b7b1573179d80dcdd620
colorama==0.4.5
commonmark==0.9.1
cycler==0.11.0
dataclasses==0.6
datasets==2.3.2
debugpy==1.6.0
#decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dill==0.3.5.1
docker-pycreds==0.4.0
editdistance==0.6.0
einops @ git+https://github.com/arogozhnikov/einops.git@d6f7910dc80c8413b7bb400f6f6ad24e3c5f5b0f
#entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
etils==0.6.0
#executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
facexlib==0.2.3
fastjsonschema==2.15.3
filelock==3.7.1
filterpy==1.4.5
flatbuffers==1.12
fonttools==4.33.3
fqdn==1.5.1
frozenlist==1.3.0
fsspec==2022.5.0
ftfy==6.1.1
future==0.18.2
fvcore==0.1.5.post20221221
gast==0.4.0
gfpgan==1.3.2
gin-config==0.5.0
gitdb==4.0.9
GitPython==3.1.27
gmx-MMPBSA==1.5.5
google-auth==2.8.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.2
grpcio==1.46.3
h5py==3.7.0
huggingface-hub==0.12.0
idna==3.3
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==4.11.4
importlib-resources==5.8.0
install==1.3.5
iopath==0.1.10
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
isoduration==20.11.0
#jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-client==8.0.3
jupyter-core==5.2.0
jupyter-events==0.6.3
jupyter-server==2.3.0
jupyter-server-fileid==0.7.0
jupyter-server-terminals==0.4.4
jupyter-server-ydoc==0.6.1
jupyter-ydoc==0.2.2
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.19.0
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
kornia==0.6.9
libclang==14.0.1
lightning-utilities==0.6.0.post0
llvmlite==0.38.1
lmdb==1.3.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mesh-tensorflow==0.1.21
mistune==0.8.4
mkl-fft==1.3.1
#mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
more-itertools==8.13.0
#mpi4py @ file:///home/conda/feedstock_root/build_artifacts/mpi4py_1649057533969/work
multidict==6.0.2
multiprocess==0.70.13
nbclassic==0.5.2
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
#nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1649847908682/work
networkx==2.8.4
nltk==3.7
notebook==6.4.12
notebook-shim==0.2.2
numba==0.55.2
#numpy @ file:///croot/numpy_and_numpy_base_1668593735768/work
oauthlib==3.2.0
omegaconf==2.0.0
opencv-python==4.1.2.30
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
#parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
#pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickle5==0.0.11
#pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.0.1
pkgutil-resolve-name==1.3.10
platformdirs==3.0.0
portalocker==2.4.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
#ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pudb==2019.2
#pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.14.1
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
#python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-json-logger==2.0.7
pytorch-lightning==1.9.1
pytorch-pretrained-biggan==0.1.1
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@3145dd4d16edaceb394838364b8e87a440f83c10
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==25.0.0
regex==2022.6.2
requests==2.28.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.4.4
rouge-score==0.0.4
rsa==4.8
s3transfer==0.6.0
sacrebleu==2.1.0
sacremoses==0.0.53
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.11.1
seqio==0.0.7
setproctitle==1.3.2
shortuuid==1.0.11
#six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.3.0
streamlit==1.10.0
t5==0.9.3
tabulate==0.8.9
# Editable Git install with no remote (taming-transformers==0.0.1)
#-e /home/yossi_gandelsman/taming-transformers
tb-nightly==2.10.0a20220615
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.9.1
tensorflow-datasets==4.6.0
tensorflow-estimator==2.9.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
tensorflow-text==2.9.0
termcolor==1.1.0
terminado==0.15.0
test-tube==0.7.5
tfds-nightly==4.6.0.dev202206170045
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==0.5.4
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.11.0+cu113
torchaudio==0.11.0+cu113
torchmetrics==0.11.1
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.0
traitlets==5.9.0
transformers==4.26.1
#typing-extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2022.1
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.13
urwid==2.1.2
validators==0.20.0
wandb==0.13.5
watchdog==2.1.9
#wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webcolors==1.12
webdataset==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
xmltodict==0.12.0
xxhash==3.0.0
y-py==0.5.9
yacs==0.1.8
yapf==0.32.0
yarl==1.7.2
ypy-websocket==0.8.2
zipp==3.8.0
